feat(apireference): Allow linking to external resources by adding a "href" attribute to <seealso /> elements
Adds support for linking to websites from within the documentation by usign a "href" attriute instead of "cref".
The value of the href attribute must be a valid uri, otherwise the seealso element is ignored.
When a <seealso /> element has both cref and href attributes the cref attribute takes precendce and the href attribute is ignored.