feat(apireference): Support external links in <see /> element
Adds support for linking to e.g. websites from using the <see /> inline
text element by specifying the target using a "href" attribute instead
of "cref".
The value of the "href" attribute must be a valid uri, otherwise the
element is ignored.
When a <see /> element has both "cref" and "href" attributes the
"cref" attribute takes precendence and the "href" attribute is ignored.