feat: Add "preset" option to adjust output files to target renderer
Adds the option to specify a "preset" to optimize the generated
documents for a specific Markdown renderer.

In the commandline version of MdDocs, the perset can be specified Using
the '--markdown-preset' parameter, for the MSBuild-integrated version,
the preset can be set using the 'MdDocsMarkdownPreset' MSBuild property.

The currently only preset other than 'Default' is 'MkDocs', which
generates Markdown suited to be included in a MkDocs projects.
At the moment the preset does not change the document structure being
generated, but uses the corresponding preset from the
'MarkdownGenerator' libary, which adjust the output for MkDocs by:

- Increasing the number of characters used to indent nested lists
- Changing the escaping of '<' and '>' to '&lt;' respectively '&gt;'