fix: Fix passing of MdSerializationOptions to MdDocuments
Because all documentation is generated through custom implementations
of IDocument the type check for MdDocument in DocumentSetExtensions did
never succeed and the MdSerializationOptions were never passed to save.

To fix this, a new interface IMarkdownDocument was added that is
implemented by all pages and allows passing a MdSerializationOptions
instance to the underlying MdDocument.