feat(commandlinehelp): Improve handling of boolean parameters
Treat boolean commandline parameters as "switch parameters"
(i.e. parameters are specified without a value, "--option" instead of "--option true").
That also implies boolean parameters cannot be required and the default value is always 'false'.

Adjust generated documentation to omit the correct syntax in the "Usage" section and
also include information that it is a switch parameter in the parameter details.

Add test class CommandLineParserTest to verify that the CommandLineParser library really treats bool parameters as switch parameters.

Closes #7