Start separating the data model from the logic to load assemblies
Introduce new types _AssemblySetDocumentation, _AssemblyDocumentation and _NamespaceDocumentation which will eventually replace the existing documentation types.
The new types are just data-types. The logic to load these from a assembly is implemented in implementations of "IDocumentationLoader" which will be the only classes tied to Mono.Cecil when the refactoring is finished.