Fix update of checked-in documentation during build of demo projects
After the build of the Demo Projects, MdDocs is started to update the Markdown output for these projects (which are checked in).
When starting MdDocs, pass in the current configuration so that the executable can be found when building "Release" (default is "Debug").

It is unclear why this was not an issue with the old build script and now surfaced after the switch to Cake.